Refactor DNS TXT record parsing for portability
The uclibc version of DNS record processing does not support TXT records
and the general processing of the DNS response into a token is fairly
straight forward so use this implementation for all clients.  This makes
the client consistent if not always right.

Added a helper function to reduce if statements for handling
error/return values.